+++ lib/inertia_rails/configuration.rb
@@ -55,6 +55,9 @@ module InertiaRails
       # When to refresh the XSRF token cookie on protected requests.
       xsrf_cookie_refresh: :always,
 
+      # Whether to convert cross-origin redirects into Inertia location responses.
+      convert_external_redirects: true,
+
       # Whether to use `<script>` element for initial page rendering instead of the `data-page` attribute.
       use_script_element_for_initial_page: false,