+++ spec/dummy/app/controllers/inertia_api_test_controller.rb
@@ -6,4 +6,8 @@ class InertiaApiTestController < ActionController::API
   def index
     render json: { ok: true }
   end
+
+  def external_redirect
+    redirect_to 'http://external-website.com/some_path', allow_other_host: true
+  end
 end